Changed how JSON import/export works.
Thre is no more copy/paste option for JSON data as
this was problematic for presentations with embedded images.

Copying/pasting 3 megs of text doesn't make the browser happy.

Exporting to JSON now prompts the browser to download the file.
Import from JSON allows the user to select the file to import from disk.