Initial integration of remoteStorage into  save/open dialogs
Open/Save As dialog allows selection of remote storage

Upon selection of remoteStorage the user may authenticate to their
remoteStorage account.

Selection of remoteStorage in the save/open modals displays the remoteStorage widget.

Includes a presentations remoteStorage module that may or may not conform to the current remoteStorage interface.

GH-8