Saver framework (auto save, save on unload, generic save)
Different from the existing save functionality in that this will apply
to all "exportable" objects.

TODO: Still need to tie this to the current
storage provider.  Storage provider should be
injected, however?  Looking it up from the registry
puts a dependency on how that lookup is performed and resolved
and requires its use with the registry..