Uploads local images to imgur in order to keep presentations small
This stems from how presentations are sync'ed to RemoteStorage.
We don't want to stick images in as base64 and keep sync'ing those
files across the network.  We're using imgur to host the images
until we can save them directly to the user's remoteStorage.

LocalStorage presents other problems that we need to figure out
how to remedy.