Save off of css based shapes before scraping and moving to SVG shapes
The problem that I ran into was coloring the CSS based shapes.

Applying an inline style such ass "border-color: red" cause all border
colors (even those that should be transparent) to be overridden.

After resolving that problem there was an issue with applying colors to
:before and :after pseudo elements since you can't apply inline styles
to pseudo elems.

After figuring out how to resolve that (by generating classes and
inserting them at the head of the document) I decided it would
be too much headache to try to manage those classes since
they would be user-generated based on the color selection tool.  I'd
have to manage saving them, not leaking them, rendering them to the
various presentations backends and so on.

The final decision is to use SVG shapes an apply inline styles to the SVG.